Added full coverage for Multi Cache session adapter
