+++ multicache/multi_cache_adapter_test.go
@@ -19,47 +19,11 @@ import (
 	"testing"
 	"time"
 
-	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/suite"
 	cacheadapters "github.com/tryvium-travels/golang-cache-adapters"
 	multicacheadapters "github.com/tryvium-travels/golang-cache-adapters/multicache"
 	testutil "github.com/tryvium-travels/golang-cache-adapters/test"
 )
-
-type mockMultiCacheAdapter struct {
-	mock.Mock
-	*multicacheadapters.MultiCacheAdapter
-}
-
-func (mca *mockMultiCacheAdapter) Get(key string, objectRef interface{}) error {
-	args := mca.Called(key, objectRef)
-
-	json.Unmarshal([]byte(testutil.TestValueJSON), &objectRef)
-
-	return args.Error(0)
-}
-
-func (mca *mockMultiCacheAdapter) Set(key string, object interface{}, newTTL *time.Duration) error {
-	args := mca.Called(key, object, newTTL)
-
-	return args.Error(0)
-}
-
-func (mca *mockMultiCacheAdapter) SetTTL(key string, newTTL time.Duration) error {
-	args := mca.Called(key, newTTL)
-
-	return args.Error(0)
-}
-
-func (mca *mockMultiCacheAdapter) Delete(key string) error {
-	args := mca.Called(key)
-
-	return args.Error(0)
-}
-
-func newmockMultiCacheAdapter() *mockMultiCacheAdapter {
-	return &mockMultiCacheAdapter{}
-}
 
 func TestMultiCacheAdapterSuite(t *testing.T) {
 	suite.Run(t, new(MultiCacheAdapterTestSuite))
